test: split backdrop coverage under file budget
